streak_junkie/activities/serializers.py
from activities.models import Activity

from rest_framework import serializers


class ActivitySerializer(serializers.ModelSerializer):
    class Meta:
        model = Activity
        fields = "__all__"
        read_only_fields = ["id"]
