requirements.txt
Django==4.0.4
djangorestframework==3.13.1
psycopg2==2.9.3

# Dev
autoflake==1.4
black==22.3.0
isort==5.10.1
